% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_PLHG}
\alias{calc_PLHG}
\title{compute the mean power analysis over the frequency band using the multitaper method}
\usage{
calc_PLHG(
  epoch,
  fs = 1000,
  sizeWindow = 3000,
  sizeSkip = 333,
  plhgTimeWindow = c(0, 20),
  baseTimeWindow = c(-30, -20)
)
}
\arguments{
\item{epoch}{Matrix or Epoch object. iEEG data matrix or Epoch object. If matrix, the row names are the electrode names and the column names are the time points}

\item{fs}{Numeric. frequency of signal iEEG acquisition}
}
\value{
A power analysis matrix
}
\description{
compute the mean power analysis over the frequency band using the multitaper method
}
\examples{
data("pt01EcoG")
epoch <- Epoch(pt01EcoG)
pt01plhg<-calc_PLHG(epoch)
}
