% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_PLHG.r
\name{nrow,PLHG-method}
\alias{nrow,PLHG-method}
\alias{ncol,PLHG-method}
\title{Get the number of rows or columns of a PLHG object}
\usage{
\S4method{nrow}{PLHG}(x)

\S4method{ncol}{PLHG}(x)
}
\arguments{
\item{x}{A PLHG object}
}
\description{
Get the number of rows or columns of a PLHG object
}
