% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ElectrodesDataPT01}
\alias{ElectrodesDataPT01}
\title{Electrodes description}
\format{
A Matrix with 84 rows (electrodes) and 3 columns
}
\source{
Fragility Multi-Center Retrospective Study
(\href{https://openneuro.org/datasets/ds003029/versions/1.0.0}{OpenNeuro})
}
\usage{
data(ElectrodesDataPT01)
}
\description{
Patient PT01 Data frame with good electrode number, classification by epileptologist as in soz and electrode name
}
\examples{
data(ElectrodesDataPT01)
}
\keyword{datasets}
