% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{heatmap_PLHG}
\alias{heatmap_PLHG}
\title{Heatmap visualization}
\usage{
heatmap_PLHG(
  plhgMaster,
  ElectrodesData,
  ieegts,
  time_window_ictal,
  subject_code,
  j
)
}
\arguments{
\item{ElectrodesData}{Electrodes data}

\item{ieegts}{Numeric. A matrix of iEEG time series x(t),
with time points as rows and electrodes names as columns}

\item{time_window_ictal}{Fragility heatmap time window around seizure onset}

\item{subject_code}{patient name}

\item{j}{seizure number}

\item{resPLHG}{Matrix result of PLHG}
}
\value{
Heatmap plot of the PLHG matrix with soz electrodes in blue in the bottom
}
\description{
Heatmap visualization
}
\examples{
data("PT01Epochm30sp30s")
data("ElectrodesDataPT01")
data("resPLHG")
ElectrodesData=ElectrodesDataPT01
time_window_ictal=c(-10,10)
subject_code='PT01'
j=1
heatmap_PLHG(plhgMaster=resPLHG[[1]], ElectrodesData=ElectrodesDataPT01,ieegts=PT01Epochm30sp30s,time_window_ictal,subject_code,j)
}
