% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPLHGHeatmap}
\alias{plotPLHGHeatmap}
\title{Visualization function (PLHG ratio matrix)}
\usage{
plotPLHGHeatmap(plhg, sozIndex = NULL)
}
\arguments{
\item{sozIndex}{Integer or string. A group of electrodes to mark as in the Seizure Onset Zone (SOZ)}

\item{EI}{EpileptogenicIndex object from \code{computeEpileptogenicIndex}}
}
\value{
A ggplot object
}
\description{
\code{plotPLHGHeatmap}: plot PLHG heatmap with electrodes marked as soz colored
}
\examples{

data("pt01EcoG")

## sozIndex is the index of the electrodes we assume are in the SOZ
sozIndex <- attr(pt01EcoG, "sozIndex")
## precomputed Epileptogenic Index object
data("pt01PLHG")

## plot the mean power heatmap
plotPLHG<-plotPLHGHeatmap(plhg=pt01PLHG,sozIndex=sozIndex)
plotPLHG
}
