% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PLHG.R
\name{votethres}
\alias{votethres}
\title{Ictal core recruitment}
\usage{
votethres(resPLHG, fs, ElectrodesData, time_window_ictal)
}
\arguments{
\item{resPLHG}{PLHG values}

\item{fs}{signal acquisition frequency}

\item{ElectrodesData}{Electrodes data names, soz/non soz epileptologist marking}

\item{time_window_ictal}{time window around seizure onset}
}
\value{
electrode classification with PLHG values increasing to 2.5SD over the mean
}
\description{
Ictal core recruitment
}
\examples{
data("ElectrodesDataPT01")
data("resPLHG")
ElectrodesData=ElectrodesDataPT01
time_window_ictal=c(-10,10)
subject_code='PT01'
j=1
fs=1000
resvotethres<-votethres(resPLHG=resPLHG,fs=fs,ElectrodesData=ElectrodesData,time_window_ictal=time_window_ictal)
}
