% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PLHG.R
\name{calc_PLHG}
\alias{calc_PLHG}
\title{Computes the Phase Locked High Gamma matrix for ictal iEEG
Implementation of PLHG as described in:}
\usage{
calc_PLHG(ieegts, sizeWindow, sizeSkip, fs, tBaseline)
}
\arguments{
\item{ieegts}{Numeric. A matrix of iEEG time series x(t),
with time points as rows and electrodes names as columns}

\item{sizeWindow}{Integer. The number of time points to use in each window}

\item{sizeSkip}{Integer.The number of time points to move the window each time}

\item{fs}{Acquisition frequency}

\item{tBaseline}{baseline start from seizure onset}
}
\value{
plhgMaster matrix of PLHG values
}
\description{
Computes the Phase Locked High Gamma matrix for ictal iEEG
Implementation of PLHG as described in:
}
\examples{
data("PT01Epochm30sp30s")
sizeWindow <- 250
sizeSkip <- 125
fs=1000
tBaseline=-30
resPLHG<-calc_PHG(ieegts = PT01Epochm30sp30s, sizeWindow=sizeWindow, sizeSkip=sizeSkip,fs=fs,tBaseline=tBaseline)
}
