% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PLHG.R
\name{calc_PLHG}
\alias{calc_PLHG}
\title{Computes the Phase Locked High Gamma matrix for ictal iEEG
Implementation of PLHG as described in:}
\usage{
calc_PLHG(
  ieegts,
  sizeWindow,
  sizeSkip,
  fs,
  tBaseline,
  time_window_ictal,
  time_window
)
}
\arguments{
\item{ieegts}{Numeric. A matrix of iEEG time series x(t),
with time points as rows and electrodes names as columns}

\item{sizeWindow}{Integer. The number of time points to use in each window}

\item{sizeSkip}{Integer.The number of time points to move the window each time}

\item{fs}{Acquisition frequency}

\item{tBaseline}{baseline duration in seconds}
}
\value{
plhgMaster matrix of PLHG values
}
\description{
Computes the Phase Locked High Gamma matrix for ictal iEEG
Implementation of PLHG as described in:
}
\examples{
data("PT01Epochm30sp30s")
sizeWindow <- 3000
sizeSkip <- 333
fs=1000
tBaseline=20
time_window_ictal=c(-10,10)
time_window=c(-30:30)
resPLHG<-calc_PHG(ieegts = PT01Epochm30sp30s, sizeWindow=sizeWindow, sizeSkip=sizeSkip,fs=fs,tBaseline=tBaseline,time_window=time_window,time_window_ictal=timewindow_ictal)
}
